# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexgohost Billing & Client Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: getMyTickets
      tags: [tickets]
      summary: Get my support tickets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create support ticket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
  /tickets/{id}/reply:
    post:
      operationId: replyToTicket
      tags: [tickets]
      summary: Reply to ticket
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketReplyRequest"
      responses:
        "201":
          description: Reply added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketMessage"
  /admin/tickets:
    get:
      operationId: getAllTickets
      tags: [tickets]
      summary: List all tickets (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
  /admin/tickets/{id}/close:
    post:
      operationId: closeTicket
      tags: [tickets]
      summary: Close ticket (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: string
        ticketNumber:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [open, closed, pending, answered]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        department:
          type: string
        lastReply:
          type: string
        messagesCount:
          type: number
        createdAt:
          type: string
      required:
        - id
        - ticketNumber
        - clientId
        - subject
        - status
        - priority
        - createdAt
    TicketDetail:
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/TicketMessage"
    TicketMessage:
      type: object
      properties:
        id:
          type: string
        ticketId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        senderRole:
          type: string
          enum: [admin, client]
        message:
          type: string
        attachments:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - ticketId
        - senderId
        - senderName
        - senderRole
        - message
        - createdAt
    CreateTicketRequest:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        department:
          type: string
      required:
        - subject
        - message
        - priority
    TicketReplyRequest:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...