# lib/api-zod/src/generated/types/ticketDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexgohost Billing & Client Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Ticket } from "./ticket";
import type { TicketMessage } from "./ticketMessage";

export type TicketDetail = Ticket & {
  messages?: TicketMessage[];
};